add folder mocks, refactory files routes
