--- data.js
@@ -1,61 +1,0 @@
-const organizations = [
-  {
-    id: 1,
-    name: "Rodrigo Fuente",
-    image: "https://drive.google.com/file/d/1ukP7z8kzFREyCbsJxzizHckgvfQ80KKt/view?usp=sharing",
-    phone: "5923457430",
-    address: "Nicaragua 767",
-    welcomeText: "Muy contento de esta aca",
-  },
-  {
-    id: 2,
-    name: "Miriam Rodriguez",
-    image: "https://drive.google.com/file/d/1PNHzyL26jzsHltv281_rc3vEKdtTg2dF/view?usp=sharing",
-    phone: "8647223019",
-    address: "Buenos Aires 2222",
-    welcomeText: "Hoy es un dia perfecto",
-  },
-  {
-    id: 3,
-    name: "Marita Gomez",
-    image: "https://drive.google.com/file/d/1PNHzyL26jzsHltv281_rc3vEKdtTg2dF/view?usp=sharing",
-    phone: "5024192311",
-    address: "Bolivar 4785",
-    welcomeText: "Gracias por ayudar",
-  },
-  {
-    id: 4,
-    name: "Maria Irola",
-    image: "https://drive.google.com/file/d/1CIsJVHyui6__A0MNAxL6hyg3Trz3ZbyW/view?usp=sharing",
-    phone: "43664552",
-    address: "Rivadavia 1111",
-    welcomeText: "Con una sonrisa todos los dias",
-  },
-  {
-    id: 5,
-    name: "Maria Garcia",
-    image: "https://drive.google.com/file/d/152GBrIzHCw2AGUpQO7jc12xQ3zaFvhK9/view?usp=sharing",
-    phone: "2632430400",
-    address: "Juncal 1986",
-    welcomeText: "Vamos por mas!!!!",
-  },
-  {
-    id: 6,
-    name: "Marco Fernandez",
-    image: "https://drive.google.com/file/d/156uGhIMSWD_iOmvZEy3-UnGFLhvLW9r5/view?usp=sharing",
-    phone: "1144552366",
-    address: "Luro 3050",
-    welcomeText: "Que bueno compartir con ustedes",
-  },
-  {
-    id: 7,
-    name: "Cecilia Mendez",
-    image: "https://drive.google.com/file/d/18Jub8i5qQnjBpuR-EsVx9Xtc0tzS2dmx/view?usp=sharing",
-    phone: "14251846",
-    address: "Gascon 7023",
-    welcomeText: "Siempre con ganas de ayudar",
-  },
-];
-
-
-module.exports = organizations;
